Update OSCA prep code as clr-transform applied to transpose of intended matrix
To perform the clr-transform, we used the mixOmics::logratio.transfo() function. We inadvertently had used the transpose of the intended matrix. We have re-run the OREML analyses and confirmed that the impact on the presented results are minior, and that the conclusions stand as in the original manuscript. We provide here the updated version of Fig.2 which directly pertains to these results. We sincerely apologise for any inconvenience caused by this error.